requirements.txt
asgiref==3.3.4
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
decorator==5.0.9
Django==3.2.4
djangorestframework==3.12.4
flake8==3.9.2
idna==2.10
ipdb==0.13.9
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
